+++ ts-client/src/amm/tests/utils/index.ts
@@ -1,6 +1,5 @@
-import { Wallet } from '@project-serum/anchor';
-import { bs58 } from '@project-serum/anchor/dist/cjs/utils/bytes';
-import { TOKEN_PROGRAM_ID, Token } from '@solana/spl-token';
+import { Wallet } from '@coral-xyz/anchor';
+import { getOrCreateAssociatedTokenAccount } from '@solana/spl-token';
 import { Connection, Keypair, LAMPORTS_PER_SOL, PublicKey } from '@solana/web3.js';
 
 export const airDropSol = async (connection: Connection, publicKey: PublicKey, amount = 1) => {
@@ -29,9 +28,7 @@ export const airDropSolIfBalanceNotEnough = async (connection: Connection, publi
 };
 
 export const getOrCreateATA = async (connection: Connection, mint: PublicKey, owner: PublicKey, payer: Keypair) => {
-  const token = new Token(connection, mint, TOKEN_PROGRAM_ID, payer);
-  const ata = await token.getOrCreateAssociatedAccountInfo(owner);
-
+  const ata = await getOrCreateAssociatedTokenAccount(connection, payer, mint, owner, true);
   return ata.address;
 };